+++ mercurysms/templates/sendSMS.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 {% load bootstrap3 %}
 {% load static %}
-{% block title %}MASS SPAM - Send SMS to list{% endblock %}
+{% block title %}Send SMS to list{% endblock %}
 
 {% block body %}
