+++ mercurysms/wsgi.py
@@ -10,7 +10,9 @@ https://docs.djangoproject.com/en/1.10/howto/deployment/wsgi/
 import os
 
 from django.core.wsgi import get_wsgi_application
+from whitenoise.django import DjangoWhiteNoise
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "sheets2sms.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "mercurysms.settings")
 
 application = get_wsgi_application()
+application = DjangoWhiteNoise(application)
