+++ requirements.txt
@@ -1,6 +1,7 @@
 appdirs==1.4.2
 Django==1.10.5
 django-bootstrap3==8.2.1
+gunicorn==19.7.1
 httplib2==0.10.3
 packaging==16.8
 pyparsing==2.1.10
@@ -9,3 +10,4 @@ pytz==2016.10
 requests==2.13.0
 six==1.10.0
 twilio==5.7.0
+whitenoise==3.3.0
